+++ Software Ray Tracer/material.h
@@ -13,7 +13,7 @@ public:
 		const ray& r_in, const hit_record& rec, color& attenuation, ray& scattered
 	) const { return false; }
 
-	virtual color emitted(double u, double v, const point3& p) const { return color(0, 0, 0); }
+	virtual color emitted(float u, float v, const point3& p) const { return color(0, 0, 0); }
 };
 
 class lambertian : public material 
@@ -101,7 +101,7 @@ public:
 	diffuse_light(shared_ptr<texture> tex) : tex(tex) {}
 	diffuse_light(const color& emit) : tex(make_shared<solid_color>(emit)) {}
 
-	color emitted(double u, double v, const point3& p) const override {
+	color emitted(float u, float v, const point3& p) const override {
 		return tex->value(u, v, p);
 	}
